feat: github workflows, gorelease and more
