+++ .github/ISSUE_TEMPLATE/Bug.md
@@ -1,0 +1,43 @@
+---
+name: Bug report
+about: Create a bug report to help us improve.
+title: ""
+labels: "bug"
+assignees: "lpmatos"
+---
+
+# Required check list
+
+- [x] I'm gonna mark the checkboxes like this.
+- [ ] I didn't find in the Gen CLI repository's issues section similar bug.
+- [ ] I understand, that the Gen CLI is Open Source and not-for-profit product.
+- [ ] This is not about third-party project, framework, package or technology.
+
+# My environment
+
+- OS (`uname -a`):
+- Golang (`go version`):
+- Gen CLI (`gen -v`):
+
+# Describe the bug
+
+<!-- A clear and concise description of what the bug is. -->
+
+# Steps to reproduce the behavior
+
+1. Go to ...
+2. Click on ...
+3. ...
+4. See error
+
+# Expected behavior
+
+<!-- A clear and concise description of what you expected to happen. -->
+
+# Screenshots
+
+<!-- If applicable, add screenshots to help explain your problem. -->
+
+# Additional context
+
+<!-- Add any other context about the problem here. -->
