+++ .github/stale.yml
@@ -1,0 +1,26 @@
+# Number of days of inactivity before an issue becomes stale
+daysUntilStale: 15
+
+# Number of days of inactivity before a stale issue is closed
+daysUntilClose: 3
+
+# Issues with these labels will never be considered stale
+exemptLabels:
+  - pinned
+  - security
+
+# Label to use when marking an issue as stale
+staleLabel: wontfix
+
+# Comment to post when marking an issue as stale. Set to `false` to disable
+markComment: >
+  👋 Hello. Is this still relevant? If so, what is blocking it? Is there 
+  anything you can do to help move it forward? 
+
+  > This issue has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs. 
+
+  Thank you for your contributions.
+
+# Comment to post when closing a stale Issue or Pull Request. Set to `false` to disable
+closeComment: >
+  ⚡️ This issue has been automatically closed because it has not had recent activity
