+++ .github/workflows/linter.yml
@@ -1,0 +1,22 @@
+name: linter
+
+on:
+  push:
+    branches:
+    - master
+  pull_request:
+    branches:
+    - master
+
+jobs:
+  linter:
+    name: golangci-lint
+    runs-on: ubuntu-latest
+    steps:
+    - name: Check out code into the Go module directory
+      uses: actions/checkout@v2
+    - name: golangci-lint
+      uses: reviewdog/action-golangci-lint@v1
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        tool_name: golangci-lint
