+++ .gitlab/merge_request_templates/General Changes.md
@@ -11,3 +11,11 @@ Please make sure your MR is ready to be merged !
 - [ ] MR targeting `master` branch.
 - [ ] MR has a green pipeline.
 - [ ] Has a green pipeline running against latest commit.
+
+## Screenshots
+
+<!-- If applicable, add screenshots to help explain your problem. -->
+
+## Additional context
+
+<!-- Add any other context about the problem here. -->
