+++ main.go
@@ -18,22 +18,8 @@ limitations under the License.
 package main
 
 import (
-	"fmt"
-
-	"github.com/common-nighthawk/go-figure"
 	"github.com/lpmatos/gen/cmd"
 )
-
-func init() {
-	title := func() {
-		figure.NewColorFigure("Gen", "", "yellow", false).Print()
-		endline := func() {
-			fmt.Println("")
-		}
-		endline()
-	}
-	title()
-}
 
 func main() {
 	cmd.Execute()
